Change the invite back
Turns out the invite wasn't the problem. Really, it seems to be
something related to caching and the interaction with the Netlify CMS
Identity script.

Opening the link in a private browsing session, and amending the
address, seems to solve the problem. Ugly, but at least there is *some*
kind of workaround.